Address review: consolidate model-tree utils into component.js [ED-22825]
- Rename findModelInDocTree → findModelInDocument
- Move addModelToParent/removeModelFromParent into document component
  utils so both JS commands and TS undoable actions share the same
  implementation
- Remove duplicated addViaModelTree/deleteViaModelTree from create.js
  and delete.js in favor of the shared utils
- Make resolve-element.ts thin wrappers delegating to the JS utils

Made-with: Cursor
