+++ packages/packages/libs/editor-elements/src/sync/resolve-element.ts
@@ -1,11 +1,5 @@
 import { getContainer } from './get-container';
-import {
-	type BackboneCollection,
-	type BackboneModel,
-	type ExtendedWindow,
-	type V1Element,
-	type V1ElementModelProps,
-} from './types';
+import { type BackboneModel, type ExtendedWindow, type V1Element, type V1ElementModelProps } from './types';
 
 function isConnected( container: V1Element | null | undefined ): container is V1Element {
 	if ( ! container ) {
@@ -35,15 +29,12 @@ export function resolveContainer( container: V1Element, id: string ): V1Element 
 	return null;
 }
 
-export function findModelInDocTree( id: string ): BackboneModel | null {
-	const extendedWindow = window as unknown as ExtendedWindow;
-	const findModelById = extendedWindow.$e?.components?.get?.( 'document' )?.utils?.findModelById;
+function getDocumentUtils() {
+	return ( window as unknown as ExtendedWindow ).$e?.components?.get?.( 'document' )?.utils;
+}
 
-	if ( ! findModelById ) {
-		return null;
-	}
-
-	return findModelById( id );
+export function findModelInDocument( id: string ): BackboneModel | null {
+	return getDocumentUtils()?.findModelById?.( id ) ?? null;
 }
 
 export function addModelToParent(
@@ -51,43 +42,9 @@ export function addModelToParent(
 	childData: V1ElementModelProps,
 	options?: { at?: number }
 ): boolean {
-	const parentModel = findModelInDocTree( parentId );
-
-	if ( ! parentModel ) {
-		return false;
-	}
-
-	const elements = parentModel.get( 'elements' ) as BackboneCollection | undefined;
-
-	if ( ! elements ) {
-		return false;
-	}
-
-	elements.add( childData, { at: options?.at, silent: true } );
-
-	return true;
+	return getDocumentUtils()?.addModelToParent?.( parentId, childData, options ) ?? false;
 }
 
 export function removeModelFromParent( parentId: string, childId: string ): boolean {
-	const parentModel = findModelInDocTree( parentId );
-
-	if ( ! parentModel ) {
-		return false;
-	}
-
-	const elements = parentModel.get( 'elements' ) as BackboneCollection | undefined;
-
-	if ( ! elements ) {
-		return false;
-	}
-
-	const child = elements.findWhere( { id: childId } );
-
-	if ( ! child ) {
-		return false;
-	}
-
-	elements.remove( child, { silent: true } );
-
-	return true;
+	return getDocumentUtils()?.removeModelFromParent?.( parentId, childId ) ?? false;
 }
